+++ mcp_fuzzer/auth/__init__.py
@@ -1,0 +1,28 @@
+from .providers import (
+    AuthProvider,
+    APIKeyAuth,
+    BasicAuth,
+    OAuthTokenAuth,
+    CustomHeaderAuth,
+    create_api_key_auth,
+    create_basic_auth,
+    create_oauth_auth,
+    create_custom_header_auth,
+)
+from .manager import AuthManager
+from .loaders import setup_auth_from_env, load_auth_config
+
+__all__ = [
+    "AuthProvider",
+    "APIKeyAuth",
+    "BasicAuth",
+    "OAuthTokenAuth",
+    "CustomHeaderAuth",
+    "create_api_key_auth",
+    "create_basic_auth",
+    "create_oauth_auth",
+    "create_custom_header_auth",
+    "AuthManager",
+    "setup_auth_from_env",
+    "load_auth_config",
+]
