+++ mcp_fuzzer/cli/__init__.py
@@ -13,6 +13,7 @@ from ..transport import create_transport
 import asyncio  # re-export for tests that patch mcp_fuzzer.cli.asyncio.run
 from rich.console import Console
 import sys  # re-export for tests that patch mcp_fuzzer.cli.sys.exit
+from ..safety_system.safety import safety_filter, disable_safety, load_safety_plugin
 
 __all__ = [
     "run_cli",
@@ -29,4 +30,7 @@ __all__ = [
     "asyncio",
     "Console",
     "sys",
+    "safety_filter",
+    "disable_safety",
+    "load_safety_plugin",
 ]
