--- mcp_fuzzer/transport.py
@@ -1,591 +1,0 @@
-"""
-Transport layer for MCP fuzzer supporting multiple protocols.
-"""
-
-import asyncio
-import sys
-import subprocess
-import os
-import signal as _signal
-import json
-import logging
-import shlex
-import uuid
-from abc import ABC, abstractmethod
-from typing import Any, Dict, List, Optional
-
-import httpx
-import websockets
-
-from .safety_system.safety import (
-    safety_filter,
-    is_safe_tool_call,
-    create_safety_response,
-    sanitize_tool_call,
-)
-
-
-class TransportProtocol(ABC):
-    """Abstract base class for transport protocols."""
-
-    @abstractmethod
-    async def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Any:
-        """Send a JSON-RPC request and return the response."""
-        pass
-
-    @abstractmethod
-    async def send_raw(self, payload: Dict[str, Any]) -> Any:
-        """
-        Send a raw JSON-RPC payload AS-IS.
-        Do not rewrite jsonrpc/id/method/params.
-        """
-        pass
-
-    @abstractmethod
-    async def send_notification(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> None:
-        """
-        Send a JSON-RPC notification (no id, no response expected).
-        """
-        pass
-
-    async def get_tools(self) -> List[Dict[str, Any]]:
-        """Get the list of tools from the server."""
-        try:
-            response = await self.send_request("tools/list")
-            logging.debug("Raw server response: %s", response)
-
-            if not isinstance(response, dict):
-                logging.warning(
-                    "Server response is not a dictionary. Got type: %s", type(response)
-                )
-                return []
-
-            if "tools" not in response:
-                logging.warning(
-                    "Server response missing 'tools' key. Keys present: %s",
-                    list(response.keys()),
-                )
-                return []
-
-            tools = response["tools"]
-            logging.info("Found %d tools from server", len(tools))
-            return tools
-
-        except Exception as e:
-            logging.exception("Failed to fetch tools from server: %s", e)
-            return []
-
-    async def call_tool(self, tool_name: str, arguments: Dict[str, Any]) -> Any:
-        """Call a specific tool with arguments."""
-        # Apply safety filtering
-        if not is_safe_tool_call(tool_name, arguments):
-            safety_filter.log_blocked_operation(
-                tool_name, arguments, "Dangerous tool call blocked in transport"
-            )
-            return create_safety_response(tool_name)
-
-        # Sanitize arguments if needed (even if not completely blocked)
-        sanitized_tool_name, sanitized_arguments = sanitize_tool_call(
-            tool_name, arguments
-        )
-
-        # Check if arguments were sanitized
-        safety_sanitized = sanitized_arguments != arguments
-
-        params = {"name": sanitized_tool_name, "arguments": sanitized_arguments}
-        result = await self.send_request("tools/call", params)
-
-        # Add safety information to the result if it was sanitized
-        if safety_sanitized and isinstance(result, dict):
-            if "_meta" not in result:
-                result["_meta"] = {}
-            result["_meta"]["safety_sanitized"] = True
-            result["_meta"]["original_arguments"] = arguments
-            result["_meta"]["sanitized_arguments"] = sanitized_arguments
-
-        return result
-
-
-class HTTPTransport(TransportProtocol):
-    """HTTP-based transport for MCP servers."""
-
-    def __init__(
-        self,
-        url: str,
-        timeout: float = 30.0,
-        auth_headers: Optional[Dict[str, str]] = None,
-    ):
-        self.url = url
-        self.timeout = timeout
-        self.headers = {
-            "Accept": "application/json, text/event-stream",
-            "Content-Type": "application/json",
-        }
-        # Add authentication headers if provided
-        if auth_headers:
-            self.headers.update(auth_headers)
-
-    async def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Any:
-        """Send a JSON-RPC request via HTTP."""
-        request_id = str(uuid.uuid4())
-        payload = {
-            "jsonrpc": "2.0",
-            "id": request_id,
-            "method": method,
-            "params": params or {},
-        }
-
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-
-            # Try to parse as JSON first
-            try:
-                data = response.json()
-            except json.JSONDecodeError:
-                # If not JSON, try to parse as SSE
-                logging.info("Response is not JSON, trying to parse as SSE")
-                for line in response.text.splitlines():
-                    if line.startswith("data:"):
-                        try:
-                            data = json.loads(line[len("data:") :].strip())
-                            break
-                        except json.JSONDecodeError:
-                            logging.error("Failed to parse SSE data line as JSON")
-                            raise
-                else:
-                    logging.error("No valid data: line found in SSE response")
-                    raise Exception("Invalid SSE response format")
-
-            if "error" in data:
-                logging.error("Server returned error: %s", data["error"])
-                raise Exception(f"Server error: {data['error']}")
-
-            # Return the result if it exists, otherwise return the full data
-            return data.get("result", data)
-
-    async def send_raw(self, payload: Dict[str, Any]) -> Any:
-        """Send a raw JSON-RPC payload via HTTP without altering fields."""
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-            try:
-                data = response.json()
-            except json.JSONDecodeError:
-                # Attempt simple SSE-like parse for data: lines
-                for line in response.text.splitlines():
-                    if line.startswith("data:"):
-                        data = json.loads(line[len("data:") :].strip())
-                        break
-                else:
-                    raise
-            if isinstance(data, dict) and "error" in data:
-                raise Exception(f"Server error: {data['error']}")
-            if isinstance(data, dict):
-                return data.get("result", data)
-            return data
-
-    async def send_notification(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> None:
-        """Send a JSON-RPC notification (no id) via HTTP."""
-        payload = {
-            "jsonrpc": "2.0",
-            "method": method,
-            "params": params or {},
-        }
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            # Fire-and-forget semantics; still POST to endpoint
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-
-
-class SSETransport(TransportProtocol):
-    """Server-Sent Events transport for MCP servers."""
-
-    def __init__(self, url: str, timeout: float = 30.0):
-        self.url = url
-        self.timeout = timeout
-        self.headers = {
-            "Accept": "text/event-stream",
-            "Content-Type": "application/json",
-        }
-
-    async def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Any:
-        """Send a JSON-RPC request via SSE."""
-        request_id = str(uuid.uuid4())
-        payload = {
-            "jsonrpc": "2.0",
-            "id": request_id,
-            "method": method,
-            "params": params or {},
-        }
-
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-
-            # Parse SSE response
-            for line in response.text.splitlines():
-                if line.startswith("data:"):
-                    try:
-                        data = json.loads(line[len("data:") :].strip())
-                        if "error" in data:
-                            logging.error("Server returned error: %s", data["error"])
-                            raise Exception(f"Server error: {data['error']}")
-                        return data.get("result", data)
-                    except json.JSONDecodeError:
-                        logging.error("Failed to parse SSE data line as JSON")
-                        continue
-
-            raise Exception("No valid SSE response received")
-
-    async def send_raw(self, payload: Dict[str, Any]) -> Any:
-        """
-        Send a raw JSON-RPC payload via SSE endpoint (POST then parse data: lines).
-        """
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-            for line in response.text.splitlines():
-                if line.startswith("data:"):
-                    try:
-                        data = json.loads(line[len("data:") :].strip())
-                    except json.JSONDecodeError:
-                        logging.error("Failed to parse SSE data line as JSON")
-                        continue
-                    if "error" in data:
-                        raise Exception(f"Server error: {data['error']}")
-                    return data.get("result", data)
-            # Some servers may respond with plain JSON
-            try:
-                data = response.json()
-                if "error" in data:
-                    raise Exception(f"Server error: {data['error']}")
-                return data.get("result", data)
-            except json.JSONDecodeError:
-                pass
-            raise Exception("No valid SSE response received")
-
-    async def send_notification(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> None:
-        payload = {
-            "jsonrpc": "2.0",
-            "method": method,
-            "params": params or {},
-        }
-        async with httpx.AsyncClient(timeout=self.timeout) as client:
-            response = await client.post(self.url, json=payload, headers=self.headers)
-            response.raise_for_status()
-
-
-class StdioTransport(TransportProtocol):
-    """Stdio-based transport for MCP servers."""
-
-    def __init__(self, command: str, timeout: float = 30.0):
-        self.command = command
-        self.timeout = timeout
-
-    async def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Any:
-        """Send a JSON-RPC request via stdio."""
-        request_id = str(uuid.uuid4())
-        payload = {
-            "jsonrpc": "2.0",
-            "id": request_id,
-            "method": method,
-            "params": params or {},
-        }
-
-        # Run the command with the request as stdin
-        process = await asyncio.create_subprocess_exec(
-            *shlex.split(self.command),
-            stdin=asyncio.subprocess.PIPE,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-            # Unix: start a new session; Windows: use CREATE_NEW_PROCESS_GROUP
-            preexec_fn=os.setsid if sys.platform != "win32" else None,
-            creationflags=(
-                subprocess.CREATE_NEW_PROCESS_GROUP if sys.platform == "win32" else 0
-            ),
-        )
-
-        stdin_data = (
-            json.dumps(payload).encode() + b"\n"
-        )  # Add newline to ensure input is sent
-        try:
-            stdout, stderr = await asyncio.wait_for(
-                process.communicate(stdin_data), timeout=self.timeout
-            )
-        except (KeyboardInterrupt, asyncio.CancelledError):
-            try:
-                if sys.platform == "win32":
-                    process.kill()
-                else:
-                    pgid = os.getpgid(process.pid)
-                    os.killpg(pgid, _signal.SIGKILL)
-            except OSError:
-                pass
-            try:
-                await process.wait()
-            except Exception:
-                pass
-            raise
-
-        if process.returncode != 0:
-            logging.error(
-                "Process failed with return code %d: %s",
-                process.returncode,
-                stderr.decode(),
-            )
-            raise Exception(f"Process failed: {stderr.decode()}")
-
-        # Parse the response
-        stdout_text = stdout.decode()
-        stderr_text = stderr.decode()
-
-        logging.debug("Process stdout: %s", stdout_text)
-        logging.debug("Process stderr: %s", stderr_text)
-
-        try:
-            # Handle multiple JSON objects in the response
-            # Split by newlines and try to parse each line as JSON
-            lines = stdout_text.strip().split("\n")
-            main_response = None
-
-            for line in lines:
-                line = line.strip()
-                if not line:
-                    continue
-
-                try:
-                    json_obj = json.loads(line)
-
-                    # Skip notification messages
-                    if json_obj.get("method") == "notifications/message":
-                        continue
-
-                    # Look for the main response (has "result" or "error" field)
-                    if "result" in json_obj or "error" in json_obj:
-                        main_response = json_obj
-                        break
-
-                except json.JSONDecodeError:
-                    # Skip lines that aren't valid JSON
-                    continue
-
-            if main_response is None:
-                # If no main response found, raise JSONDecodeError
-                raise json.JSONDecodeError("No main response found", stdout_text, 0)
-            else:
-                data = main_response
-
-            if "error" in data:
-                logging.error("Server returned error: %s", data["error"])
-                raise Exception(f"Server error: {data['error']}")
-            return data.get("result", data)
-        except json.JSONDecodeError:
-            logging.error("Failed to parse response as JSON: %s", stdout_text)
-            logging.error("Process stderr: %s", stderr_text)
-            raise
-
-    async def send_raw(self, payload: Dict[str, Any]) -> Any:
-        """Send a raw JSON-RPC payload via stdio."""
-        process = await asyncio.create_subprocess_exec(
-            *shlex.split(self.command),
-            stdin=asyncio.subprocess.PIPE,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-            preexec_fn=os.setsid if sys.platform != "win32" else None,
-            creationflags=(
-                subprocess.CREATE_NEW_PROCESS_GROUP if sys.platform == "win32" else 0
-            ),
-        )
-
-        stdin_data = json.dumps(payload).encode() + b"\n"
-        try:
-            stdout, stderr = await asyncio.wait_for(
-                process.communicate(stdin_data), timeout=self.timeout
-            )
-        except (KeyboardInterrupt, asyncio.CancelledError):
-            try:
-                if sys.platform == "win32":
-                    process.kill()
-                else:
-                    pgid = os.getpgid(process.pid)
-                    os.killpg(pgid, _signal.SIGKILL)
-            except OSError:
-                pass
-            try:
-                await process.wait()
-            except Exception:
-                pass
-            raise
-
-        if process.returncode != 0:
-            logging.error(
-                "Process failed with return code %d: %s",
-                process.returncode,
-                stderr.decode(),
-            )
-            raise Exception(f"Process failed: {stderr.decode()}")
-
-        stdout_text = stdout.decode()
-        try:
-            lines = stdout_text.strip().split("\n")
-            main_response = None
-            for line in lines:
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    json_obj = json.loads(line)
-                except json.JSONDecodeError:
-                    continue
-                if isinstance(json_obj, dict) and (
-                    "result" in json_obj or "error" in json_obj
-                ):
-                    main_response = json_obj
-                    break
-            if main_response is None:
-                raise json.JSONDecodeError("No main response found", stdout_text, 0)
-            if "error" in main_response:
-                raise Exception(f"Server error: {main_response['error']}")
-            return main_response.get("result", main_response)
-        except json.JSONDecodeError:
-            logging.error("Failed to parse response as JSON: %s", stdout_text)
-            raise
-
-    async def send_notification(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> None:
-        payload = {
-            "jsonrpc": "2.0",
-            "method": method,
-            "params": params or {},
-        }
-        process = await asyncio.create_subprocess_exec(
-            *shlex.split(self.command),
-            stdin=asyncio.subprocess.PIPE,
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.PIPE,
-            preexec_fn=os.setsid if sys.platform != "win32" else None,
-            creationflags=(
-                subprocess.CREATE_NEW_PROCESS_GROUP if sys.platform == "win32" else 0
-            ),
-        )
-        stdin_data = json.dumps(payload).encode() + b"\n"
-        try:
-            stdout, stderr = await asyncio.wait_for(
-                process.communicate(stdin_data), timeout=self.timeout
-            )
-        except (KeyboardInterrupt, asyncio.CancelledError):
-            try:
-                if sys.platform == "win32":
-                    process.kill()
-                else:
-                    pgid = os.getpgid(process.pid)
-                    os.killpg(pgid, _signal.SIGKILL)
-            except OSError:
-                pass
-            try:
-                await process.wait()
-            except Exception:
-                pass
-            raise
-        if process.returncode != 0:
-            logging.error(
-                "Notification subprocess failed with return code %d: %s",
-                process.returncode,
-                (stderr or b"").decode(),
-            )
-            raise Exception(f"Notification process failed: {(stderr or b'').decode()}")
-
-
-class WebSocketTransport(TransportProtocol):
-    """WebSocket-based transport for MCP servers."""
-
-    def __init__(self, url: str, timeout: float = 30.0):
-        self.url = url
-        self.timeout = timeout
-
-    async def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Any:
-        """Send a JSON-RPC request via WebSocket."""
-        request_id = str(uuid.uuid4())
-        payload = {
-            "jsonrpc": "2.0",
-            "id": request_id,
-            "method": method,
-            "params": params or {},
-        }
-
-        async with websockets.connect(self.url) as websocket:
-            await websocket.send(json.dumps(payload))
-
-            try:
-                response = await asyncio.wait_for(
-                    websocket.recv(), timeout=self.timeout
-                )
-                data = json.loads(response)
-
-                if "error" in data:
-                    logging.error("Server returned error: %s", data["error"])
-                    raise Exception(f"Server error: {data['error']}")
-
-                return data.get("result", data)
-            except asyncio.TimeoutError:
-                raise Exception("WebSocket request timed out")
-
-    async def send_raw(self, payload: Dict[str, Any]) -> Any:
-        """Send a raw JSON-RPC payload via WebSocket."""
-        async with websockets.connect(self.url) as websocket:
-            await websocket.send(json.dumps(payload))
-            try:
-                response = await asyncio.wait_for(
-                    websocket.recv(), timeout=self.timeout
-                )
-                data = json.loads(response)
-                if isinstance(data, dict) and "error" in data:
-                    raise Exception(f"Server error: {data['error']}")
-                if isinstance(data, dict):
-                    return data.get("result", data)
-                return data
-            except asyncio.TimeoutError:
-                raise Exception("WebSocket request timed out")
-
-    async def send_notification(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> None:
-        payload = {
-            "jsonrpc": "2.0",
-            "method": method,
-            "params": params or {},
-        }
-        async with websockets.connect(self.url) as websocket:
-            await websocket.send(json.dumps(payload))
-
-
-def create_transport(protocol: str, endpoint: str, **kwargs) -> TransportProtocol:
-    """Factory function to create the appropriate transport based on protocol."""
-    if protocol == "http":
-        return HTTPTransport(endpoint, **kwargs)
-    elif protocol == "sse":
-        return SSETransport(endpoint, **kwargs)
-    elif protocol == "stdio":
-        return StdioTransport(endpoint, **kwargs)
-    elif protocol == "websocket":
-        return WebSocketTransport(endpoint, **kwargs)
-    else:
-        raise ValueError(f"Unsupported protocol: {protocol}")
