+++ mcp_fuzzer/transport/__init__.py
@@ -1,0 +1,24 @@
+from .base import TransportProtocol
+from .http import HTTPTransport
+from .sse import SSETransport
+from .stdio import StdioTransport
+from .websocket import WebSocketTransport
+from .factory import create_transport
+
+# Back-compat for tests that patch mcp_fuzzer.transport.httpx / asyncio / websockets
+import httpx  # noqa: F401
+import asyncio  # noqa: F401
+import websockets  # noqa: F401
+
+__all__ = [
+    "TransportProtocol",
+    "HTTPTransport",
+    "SSETransport",
+    "StdioTransport",
+    "WebSocketTransport",
+    "create_transport",
+    # modules exposed for test patching
+    "httpx",
+    "asyncio",
+    "websockets",
+]
